refactor(test): move connector helpers to db-test-seeders and db-test-assertions
Move 5 DB-direct connector test helpers from api-test-helpers/connectors.ts
to their proper locations:
- 3 seeders to db-test-seeders/connectors.ts (with @why-db-direct annotations)
- 2 assertion helpers to db-test-assertions/connectors.ts

API-based createTestConnector and its private helpers remain in
api-test-helpers/connectors.ts. After this change, the file has zero
db/schema imports.

Closes #9359
