+++ turbo/apps/web/src/__tests__/db-test-assertions/connectors.ts
@@ -1,0 +1,60 @@
+import { and, eq } from "drizzle-orm";
+import { initServices } from "../../lib/init-services";
+import { secrets } from "../../db/schema/secret";
+import { connectors } from "../../db/schema/connector";
+import { decryptSecretValue } from "../../lib/shared/crypto/secrets-encryption";
+
+// ---------------------------------------------------------------------------
+// Read-only assertion helpers for connector test verification.
+// ---------------------------------------------------------------------------
+
+/**
+ * Find and decrypt a connector secret token from the database.
+ * Used for verifying that the correct token was stored during connector
+ * OAuth flow.
+ */
+export async function findTestConnectorSecret(
+  orgId: string,
+  secretName: string,
+  type: "connector" | "user" = "connector",
+): Promise<string | undefined> {
+  initServices();
+  const [storedSecret] = await globalThis.services.db
+    .select()
+    .from(secrets)
+    .where(
+      and(
+        eq(secrets.orgId, orgId),
+        eq(secrets.name, secretName),
+        eq(secrets.type, type),
+      ),
+    )
+    .limit(1);
+
+  if (!storedSecret) return undefined;
+
+  return decryptSecretValue(
+    storedSecret.encryptedValue,
+    globalThis.services.env.SECRETS_ENCRYPTION_KEY,
+  );
+}
+
+/**
+ * Get the tokenExpiresAt timestamp for a connector.
+ * Used for verifying that token expiry was correctly stored during OAuth
+ * flow.
+ */
+export async function findTestConnectorTokenExpiresAt(
+  orgId: string,
+  type: string,
+): Promise<Date | null | undefined> {
+  initServices();
+  const [row] = await globalThis.services.db
+    .select({ tokenExpiresAt: connectors.tokenExpiresAt })
+    .from(connectors)
+    .where(and(eq(connectors.orgId, orgId), eq(connectors.type, type)))
+    .limit(1);
+
+  if (!row) return undefined;
+  return row.tokenExpiresAt;
+}
